sytle: 패키지 구조를 도메인형 구조로 변경
