+++ courseX-backend/src/main/java/com/acc/courseX/common/code/BaseCode.java
@@ -1,0 +1,9 @@
+package com.acc.courseX.common.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseCode {
+  HttpStatus getStatus();
+
+  String getMessage();
+}
