+++ courseX-backend/src/main/java/com/acc/courseX/common/code/FailureCode.java
@@ -1,0 +1,3 @@
+package com.acc.courseX.common.code;
+
+public interface FailureCode extends BaseCode {}
