+++ courseX-backend/src/main/java/com/acc/courseX/common/code/SuccessCode.java
@@ -1,0 +1,3 @@
+package com.acc.courseX.common.code;
+
+public interface SuccessCode extends BaseCode {}
