--- courseX-backend/src/main/java/com/acc/courseX/controller/CourseApi.java
@@ -1,7 +1,0 @@
-package com.acc.courseX.controller;
-
-import org.springframework.http.ResponseEntity;
-
-public interface CourseApi {
-  ResponseEntity<?> getCourses(String code);
-}
