+++ courseX-backend/src/main/java/com/acc/courseX/course/controller/CourseApi.java
@@ -1,0 +1,7 @@
+package com.acc.courseX.course.controller;
+
+import org.springframework.http.ResponseEntity;
+
+public interface CourseApi {
+  ResponseEntity<?> getCourses(String code);
+}
