+++ courseX-backend/src/main/java/com/acc/courseX/course/entity/CourseType.java
@@ -1,0 +1,6 @@
+package com.acc.courseX.course.entity;
+
+public enum CourseType {
+  MAJOR,
+  GENERAL
+}
