+++ courseX-backend/src/main/java/com/acc/courseX/course/service/CourseService.java
@@ -1,0 +1,9 @@
+package com.acc.courseX.course.service;
+
+import java.util.List;
+
+import com.acc.courseX.course.dto.CourseResponse;
+
+public interface CourseService {
+  List<CourseResponse> getCourses(String code);
+}
