+++ courseX-backend/src/main/java/com/acc/courseX/enrollment/entity/EnrollmentStatus.java
@@ -1,0 +1,6 @@
+package com.acc.courseX.enrollment.entity;
+
+public enum EnrollmentStatus {
+  ENROLLED,
+  CANCELLED
+}
