--- courseX-backend/src/main/java/com/acc/courseX/entity/BaseTime.java
@@ -1,29 +1,0 @@
-package com.acc.courseX.entity;
-
-import static lombok.AccessLevel.PROTECTED;
-
-import java.time.LocalDateTime;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@NoArgsConstructor(access = PROTECTED)
-public abstract class BaseTime {
-  @Column(updatable = false, name = "created_at")
-  @CreatedDate
-  private LocalDateTime createdAt;
-
-  @Column(name = "updated_at")
-  @LastModifiedDate
-  private LocalDateTime updatedAt;
-}
