--- courseX-backend/src/main/java/com/acc/courseX/entity/CourseSchedule.java
@@ -1,48 +1,0 @@
-package com.acc.courseX.entity;
-
-import java.time.LocalTime;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-import com.acc.courseX.entity.enums.Weekday;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "course_schedules")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CourseSchedule {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "course_id", nullable = false)
-  private Course course;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private Weekday weekday;
-
-  @Column(nullable = false)
-  private LocalTime startTime;
-
-  @Column(nullable = false)
-  private LocalTime endTime;
-
-  @Override
-  public String toString() {
-    return weekday.getDisplayName() + ": " + startTime + " - " + endTime;
-  }
-}
