--- courseX-backend/src/main/java/com/acc/courseX/entity/Major.java
@@ -1,23 +1,0 @@
-package com.acc.courseX.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "majors")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Major {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false)
-  private String name;
-}
