--- courseX-backend/src/main/java/com/acc/courseX/entity/User.java
@@ -1,46 +1,0 @@
-package com.acc.courseX.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-import com.acc.courseX.entity.enums.Role;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "users")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class User extends BaseTime {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false, unique = true)
-  private String affiliationId;
-
-  @Column(nullable = false)
-  private String password;
-
-  @Column(nullable = false)
-  private String name;
-
-  @Enumerated(value = EnumType.STRING)
-  @Column(nullable = false)
-  private Role role = Role.STUDENT;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "major_id", nullable = false)
-  private Major major;
-}
