--- courseX-backend/src/main/java/com/acc/courseX/entity/enums/CourseType.java
@@ -1,6 +1,0 @@
-package com.acc.courseX.entity.enums;
-
-public enum CourseType {
-  MAJOR,
-  GENERAL
-}
