--- courseX-backend/src/main/java/com/acc/courseX/entity/enums/EnrollmentStatus.java
@@ -1,6 +1,0 @@
-package com.acc.courseX.entity.enums;
-
-public enum EnrollmentStatus {
-  ENROLLED,
-  CANCELLED
-}
