--- courseX-backend/src/main/java/com/acc/courseX/entity/enums/Role.java
@@ -1,6 +1,0 @@
-package com.acc.courseX.entity.enums;
-
-public enum Role {
-  STUDENT,
-  PROFESSOR
-}
