--- courseX-backend/src/main/java/com/acc/courseX/repository/CourseRepository.java
@@ -1,11 +1,0 @@
-package com.acc.courseX.repository;
-
-import java.util.List;
-
-import com.acc.courseX.entity.Course;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CourseRepository extends JpaRepository<Course, Long> {
-  List<Course> findByCode(String code);
-}
