--- courseX-backend/src/main/java/com/acc/courseX/service/CourseService.java
@@ -1,9 +1,0 @@
-package com.acc.courseX.service;
-
-import java.util.List;
-
-import com.acc.courseX.dto.CourseResponse;
-
-public interface CourseService {
-  List<CourseResponse> getCourses(String code);
-}
