--- courseX-backend/src/main/java/com/acc/courseX/support/code/BaseCode.java
@@ -1,9 +1,0 @@
-package com.acc.courseX.support.code;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseCode {
-  HttpStatus getStatus();
-
-  String getMessage();
-}
