--- courseX-backend/src/main/java/com/acc/courseX/support/code/FailureCode.java
@@ -1,3 +1,0 @@
-package com.acc.courseX.support.code;
-
-public interface FailureCode extends BaseCode {}
