--- courseX-backend/src/main/java/com/acc/courseX/support/code/SuccessCode.java
@@ -1,3 +1,0 @@
-package com.acc.courseX.support.code;
-
-public interface SuccessCode extends BaseCode {}
