--- courseX-backend/src/main/java/com/acc/courseX/support/exception/BaseException.java
@@ -1,18 +1,0 @@
-package com.acc.courseX.support.exception;
-
-import com.acc.courseX.support.code.FailureCode;
-
-public abstract class BaseException extends RuntimeException {
-
-  private final FailureCode failure;
-
-  public BaseException(final FailureCode failure) {
-    super(failure.getMessage());
-    this.failure = failure;
-  }
-
-  public FailureCode getError() {
-    return this.failure;
-  }
-  ;
-}
