+++ courseX-backend/src/main/java/com/acc/courseX/user/entity/Role.java
@@ -1,0 +1,6 @@
+package com.acc.courseX.user.entity;
+
+public enum Role {
+  STUDENT,
+  PROFESSOR
+}
